added /api/types/types.ts for common types. moved all dtos to /api/types/api/...
