--- lib/api/updatePrivilegeLevel.dto.ts
@@ -1,10 +1,0 @@
-import { PrismaClient } from '@prisma/client';
-
-export type Context = {
-  prisma: PrismaClient;
-};
-
-export type RequestParameters = {
-  targetEmail: string;
-  targetPrivilegeLevel: string;
-};
