+++ lib/types/api/createNewsRequest.dto.ts
