+++ lib/types/api/updatePrivilegeLevel.dto.ts
@@ -1,0 +1,4 @@
+export type RequestParameters = {
+  targetEmail: string;
+  targetPrivilegeLevel: string;
+};
