+++ lib/types/next-auth.d.ts
