+++ lib/types/types.ts
@@ -1,0 +1,5 @@
+import { PrismaClient } from "@prisma/client";
+
+export type Context = {
+  prisma: PrismaClient;
+};
